Clean up functional tests
