+++ scico/test/functional/prox.py
