+++ scico/test/functional/test_denoiser.py
@@ -1,0 +1,46 @@
+import numpy as np
+
+import pytest
+
+from scico import denoiser, functional
+from scico.denoiser import have_bm3d
+from scico.random import randn
+
+
+@pytest.mark.skipif(not have_bm3d, reason="bm3d package not installed")
+class TestBM3D:
+    def setup(self):
+        key = None
+        self.x_gry, key = randn((32, 33), key=key, dtype=np.float32)
+        self.x_rgb, key = randn((33, 34, 3), key=key, dtype=np.float32)
+        self.f_gry = functional.BM3D()
+        self.f_rgb = functional.BM3D(is_rgb=True)
+
+    def test_gry(self):
+        y0 = self.f_gry.prox(self.x_gry, 1.0)
+        y1 = denoiser.bm3d(self.x_gry, 1.0)
+        np.testing.assert_allclose(y0, y1, rtol=1e-5)
+
+    def test_rgb(self):
+        y0 = self.f_rgb.prox(self.x_rgb, 1.0)
+        y1 = denoiser.bm3d(self.x_rgb, 1.0, is_rgb=True)
+        np.testing.assert_allclose(y0, y1, rtol=1e-5)
+
+
+class TestDnCNN:
+    def setup(self):
+        key = None
+        self.x_sngchn, key = randn((32, 33), key=key, dtype=np.float32)
+        self.x_mltchn, key = randn((33, 34, 5), key=key, dtype=np.float32)
+        self.dncnn = denoiser.DnCNN()
+        self.f = functional.DnCNN()
+
+    def test_sngchn(self):
+        y0 = self.f.prox(self.x_sngchn, 1.0)
+        y1 = self.dncnn(self.x_sngchn)
+        np.testing.assert_allclose(y0, y1, rtol=1e-5)
+
+    def test_mltchn(self):
+        y0 = self.f.prox(self.x_mltchn, 1.0)
+        y1 = self.dncnn(self.x_mltchn)
+        np.testing.assert_allclose(y0, y1, rtol=1e-5)
