+++ scico/test/functional/test_loss.py
