+++ scico/test/functional/test_misc.py
@@ -1,0 +1,67 @@
+import numpy as np
+
+import jax
+
+import pytest
+
+import scico.numpy as snp
+from scico import functional
+
+
+class TestCheckAttrs:
+    # Ensure that the has_eval, has_prox attrs are overridden
+    # and set to True/False in the Functional subclasses.
+
+    # Generate a list of all functionals in scico.functionals that we will check
+    ignore = [functional.Functional, functional.ScaledFunctional, functional.SeparableFunctional]
+    to_check = []
+    for name, cls in functional.__dict__.items():
+        if isinstance(cls, type):
+            if issubclass(cls, functional.Functional):
+                if cls not in ignore:
+                    to_check.append(cls)
+
+    @pytest.mark.parametrize("cls", to_check)
+    def test_has_eval(self, cls):
+        assert isinstance(cls.has_eval, bool)
+
+    @pytest.mark.parametrize("cls", to_check)
+    def test_has_prox(self, cls):
+        assert isinstance(cls.has_prox, bool)
+
+
+def test_functional_sum():
+    x = np.random.randn(4, 4)
+    f0 = functional.L1Norm()
+    f1 = 2.0 * functional.L2Norm()
+    f = f0 + f1
+    assert f(x) == f0(x) + f1(x)
+    with pytest.raises(TypeError):
+        f = f0 + 2.0
+
+
+def test_scalar_vmap():
+    x = np.random.randn(4, 4)
+    f = functional.L1Norm()
+
+    def foo(c):
+        return (c * f)(x)
+
+    c_list = [1.0, 2.0, 3.0]
+    non_vmap = np.array([foo(c) for c in c_list])
+
+    vmapped = jax.vmap(foo)(snp.array(c_list))
+    np.testing.assert_allclose(non_vmap, vmapped)
+
+
+def test_scalar_pmap():
+    x = np.random.randn(4, 4)
+    f = functional.L1Norm()
+
+    def foo(c):
+        return (c * f)(x)
+
+    c_list = np.random.randn(jax.device_count())
+    non_pmap = np.array([foo(c) for c in c_list])
+    pmapped = jax.pmap(foo)(c_list)
+    np.testing.assert_allclose(non_pmap, pmapped)
